+++ src/client/sync/uploads.rs
@@ -6,7 +6,7 @@ use failure::{Error, err_msg, ResultExt};
 use ssh_keys::PrivateKey;
 
 use ciruela::blocks::ThreadedBlockReader;
-use ciruela::index::InMemoryIndexes;
+use ciruela::index::{InMemoryIndexes, ImageId};
 use ciruela::signature::{SignedUpload, sign_upload};
 use ciruela::VPath;
 
@@ -19,16 +19,31 @@ pub enum Upload {
     Append(SignedUpload),
     WeakAppend(SignedUpload),
     Replace(SignedUpload),
+    ReplaceIfMatches(SignedUpload, ImageId),
 }
 
 fn split(cli: &str) -> Result<(PathBuf, VPath), Error> {
     let mut pair = cli.split(':');
     let src = PathBuf::from(pair.next().unwrap());
-    // TODO(tailhook) this may crash
-    let dest = VPath::from(pair.next()
-        .ok_or_else(|| err_msg(format!("No destination directory specified, \
-            must be `source:/dir/dest`, got {:?} instead", cli)))?);
+    let dest = pair.next().and_then(|x| VPath::try_from(x).ok())
+        .ok_or_else(|| err_msg(format!("Destination directory is invalid, \
+            must be `source:/dir/dest`, got {:?} instead", cli)))?;
     return Ok((src, dest));
+}
+
+fn split_replace(cli: &str)
+    -> Result<(PathBuf, VPath, Option<ImageId>), Error>
+{
+    let mut triple = cli.split(':');
+    let src = PathBuf::from(triple.next().unwrap());
+    // TODO(tailhook) this may crash
+    let dest = triple.next().and_then(|x| VPath::try_from(x).ok())
+        .ok_or_else(|| err_msg(format!("Destination directory is invalid, \
+            must be `source:/dir/dest`, got {:?} instead", cli)))?;
+    let image_id = triple.next().map(|x| x.parse())
+        .map_or(Ok(None), |v| v.map(Some))
+        .map_err(|e| format_err!("error parsing image_id: {}", e))?;
+    return Ok((src, dest, image_id));
 }
 
 fn scan(dir: &Path, threads: usize) -> Result<Vec<u8>, Error> {
@@ -73,13 +88,17 @@ pub(in sync) fn prepare(opts: &SyncOptions, keys: &Vec<PrivateKey>,
     }
 
     for dir in &opts.replace {
-        let (src, dest) = split(dir)?;
+        let (src, dest, old_image) = split_replace(dir)?;
         let index_buf = scan(&src, gopt.threads)?;
         let image_id = indexes.register_index(&index_buf)?;
         blocks.register_dir(&src, &index_buf)?;
 
         let upload = sign_upload(&dest, &image_id, timestamp, &keys);
-        result.push(Upload::Replace(upload));
+        if let Some(old) = old_image {
+            result.push(Upload::ReplaceIfMatches(upload, old));
+        } else {
+            result.push(Upload::Replace(upload));
+        }
     }
 
     return Ok(result);